Moved signal tests into separate file
