+++ rdmo/views/tests/test_viewset_view.py
@@ -1,13 +1,7 @@
 import xml.etree.ElementTree as et
 
-import itertools
 import pytest
-from django.contrib.auth.models import Group
-from django.contrib.sites.models import Site
 from django.urls import reverse
-
-from rdmo.projects.models import Project
-from rdmo.questions.models import Catalog
 
 from ..models import View
 
@@ -18,12 +12,6 @@ users = (
     ('api', 'api'),
     ('anonymous', None),
 )
-
-view_project_test_users = (
-    ('editor', 'editor'),
-    ('api', 'api'),
-)
-
 
 status_map = {
     'list': {
@@ -52,38 +40,6 @@ urlnames = {
     'copy': 'v1-views:view-copy'
 }
 
-view_update_tests = [
-    # tuples of: view_id, sites, catalogs, groups, project_id, project_exists
-    ('3', [],       [],     [], '10', True),
-    ('3', [2],      [],     [], '10', False),
-    ('3', [1,2,3],  [],     [], '10', True),
-    ('3', [],       [2],    [], '10', False),
-    ('3', [2],      [2],    [], '10', False),
-    ('3', [1,2,3],  [2],    [], '10', False),
-    ('3', [],       [1,2],  [], '10', True),
-    ('3', [2],      [1,2],  [], '10', False),
-    ('3', [1,2,3],  [1,2],  [], '10', True),
-
-    ('3', [],       [],     [1], '10', False),
-    ('3', [2],      [],     [1], '10', False),
-    ('3', [1,2,3],  [],     [1], '10', False),
-    ('3', [],       [2],    [1], '10', False),
-    ('3', [2],      [2],    [1], '10', False),
-    ('3', [1,2,3],  [2],    [1], '10', False),
-    ('3', [],       [1,2],  [1], '10', False),
-    ('3', [2],      [1,2],  [1], '10', False),
-    ('3', [1,2,3],  [1,2],  [1], '10', False),
-
-    ('3', [],       [],     [1,2,3,4], '10', True),
-    ('3', [2],      [],     [1,2,3,4], '10', False),
-    ('3', [1,2,3],  [],     [1,2,3,4], '10', True),
-    ('3', [],       [2],    [1,2,3,4], '10', False),
-    ('3', [2],      [2],    [1,2,3,4], '10', False),
-    ('3', [1,2,3],  [2],    [1,2,3,4], '10', False),
-    ('3', [],       [1,2],  [1,2,3,4], '10', True),
-    ('3', [2],      [1,2],  [1,2,3,4], '10', False),
-    ('3', [1,2,3],  [1,2],  [1,2,3,4], '10', True)
-]
 
 @pytest.mark.parametrize('username,password', users)
 def test_list(db, client, username, password):
@@ -169,25 +125,6 @@ def test_update(db, client, username, password):
         }
         response = client.put(url, data, content_type='application/json')
         assert response.status_code == status_map['update'][username], response.json()
-
-@pytest.mark.parametrize('view_id,sites,catalogs,groups,project_id,project_exists', view_update_tests)
-def test_update_projects(db, view_id, sites, catalogs, groups, project_id, project_exists):
-    view = View.objects.get(pk=view_id)
-    project = Project.objects.get(pk=project_id)
-
-    view.sites.set(Site.objects.filter(pk__in=sites))
-    view.catalogs.set(Catalog.objects.filter(pk__in=catalogs))
-    view.groups.set(Group.objects.filter(pk__in=groups))
-
-    assert sorted(list(itertools.chain.from_iterable(view.sites.all().values_list('pk')))) == sites
-    assert sorted(list(itertools.chain.from_iterable(view.catalogs.all().values_list('pk')))) == catalogs
-    assert sorted(list(itertools.chain.from_iterable(view.groups.all().values_list('pk')))) == groups
-
-    if not project_exists:
-        with pytest.raises(Project.DoesNotExist):
-            Project.objects.filter(views=view).get(pk=project_id)
-    else:
-        assert Project.objects.filter(views=view).get(pk=project_id)
 
 @pytest.mark.parametrize('username,password', users)
 def test_delete(db, client, username, password):
